+++ api/Repositories/BidRepo/BidRepository.cs
@@ -17,7 +17,7 @@ public class BidRepository : IBidRepository // Implementacja IBidRepository poka
     {
         return await _context.Bids.ToListAsync();// Zapisuje to wszystko do listy i zwraca ta liste.
     }
-    public async Task<Bid?> GetByIdAsync(int id) // Metoda GetByIdAsync przyjmuje jako parametr id typu int, ktore jest ide ntyfikatorem oferty (bid) w bazie danych. Metoda zwraca obiekt Bid lub null, jesli oferta o podanym id nie istnieje.
+    public async Task<Bid?> GetByIdAsync(long id) // Metoda GetByIdAsync przyjmuje jako parametr id typu int, ktore jest ide ntyfikatorem oferty (bid) w bazie danych. Metoda zwraca obiekt Bid lub null, jesli oferta o podanym id nie istnieje.
     {
         return await _context.Bids.FindAsync(id); // Uzywa metody FindAsync() ktora szuka w bazie danych oferty o podanym id i zwraca ja, jesli znajdzie, lub null, jesli nie znajdzie.
     }
@@ -33,7 +33,7 @@ public class BidRepository : IBidRepository // Implementacja IBidRepository poka
         await _context.SaveChangesAsync();// Zapisuje wszystkie zmiany w bazie danych.
         return bid;
     }
-    public async Task DeleteBidAsync(int id) // Przyjmuje id 
+    public async Task DeleteBidAsync(long id) // Przyjmuje id 
     {
         var Bid = await _context.Bids.FindAsync(id); // Szuka w bazie danych id i przypisuje to do zmiennej Bid.
         if (Bid == null) // Sprawdza czy jest Bid jest null
