+++ api/Services/AuctionS/IAuctionService.cs
@@ -5,10 +5,10 @@ namespace api.Services.AuctionS
     public interface IAuctionService
     {
         Task<IEnumerable<Auction>> GetAllAsync();
-        Task<Auction?> GetByIdAsync(int id);
+        Task<Auction?> GetByIdAsync(long id);
         Task<Auction> AddAuctionAsync(Auction auction);
         Task<Auction> UpdateAuctionAsync(Auction auction);
-        Task DeleteAuctionAsync(int id);
+        Task DeleteAuctionAsync(long id);
 
     }
 }
